requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
websockets==12.0
requests==2.32.3
python-dotenv==1.0.1
twilio==9.3.0
deepgram-sdk==3.2.7
TTS==0.22.0
soundfile==0.12.1
librosa==0.10.2.post1
numpy==1.26.4
scipy==1.13.1
pydantic==2.7.3
